Add a number of utils
