+++ src/scripts/pages/all.ts
@@ -2,11 +2,13 @@ import { Logger } from "../utils/logger";
 import { Setting, updateSettings } from "../utils/settings";
 import { getBrowser } from "../utils/dom";
 
-function load() {
-    Logger.log(`Loaded Schoology Plus version ${chrome.runtime.getManifest().version}${getBrowser() != "Chrome" || chrome.runtime.getManifest().update_url ? '' : ' (development version)'}`);
+import browser from "webextension-polyfill";
+
+async function load() {
+    Logger.log(`Loaded Schoology Plus version ${browser.runtime.getManifest().version}${getBrowser() != "Chrome" || (browser.runtime.getManifest() as any).update_url ? '' : ' (development version)'}`);
     document.documentElement.setAttribute("page", location.pathname);
 
-    updateSettings();
+    await updateSettings();
 
     new Setting(
         "defaultDomain",
